+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.2.2 ()
+* [#50](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/52): Can use relative path for test result file path
+* [#45](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/45): Run test(s) in context
 * [#46](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/48): Changing pathForResultFile does not require code reload
 * [#43](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/43): Go to test code from tree